pyproject.toml
[project]
name = "telegramtranslator"
version = "0.1.0"
description = ""
authors = [
    {name = "BuildTools",email = "unconfigured@null.spigotmc.org"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "aiogram (>=3.20.0.post0,<4.0.0)",
    "langdetect (>=1.0.9,<2.0.0)",
    "transformers (>=4.52.4,<5.0.0)",
    "torch (>=2.7.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "lingua-language-detector (>=2.1.1,<3.0.0)",
    "sentencepiece (>=0.2.0,<0.3.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
lingua-language-detector = "^2.1.1"
uvicorn = "^0.34.3"
pytest-asyncio = "^1.0.0"
pre-commit = "^4.2.0"
isort = "^6.0.1"
black = "^25.1.0"

[tool.pytest.ini_options]
pythonpath = ["."]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"      # гарантирует, что isort не спорит с black
line_length = 88
known_first_party = ["app"]
